fix(stream): read the END reply's count instead of the result code
Every CAS reply starts with a result code, and the stream END reply puts the
binding's count after it -- rows for COPY, bytes for a value stream, 64-bit so a
4GB LOB value fits. streamEnd () returned the result code and never read the
count, so it answered 0 on every successful COPY while its own javadoc promised
the number of rows loaded. The eight bytes stayed unread in the buffer.

Read the count and return it. long, because the count is 64-bit on the wire and
truncating it here would put back the limit the width was chosen to remove.

The C driver took this fix already; this is the same one on the Java side.

+++ src/jdbc/cubrid/jdbc/driver/CUBRIDConnection.java
@@ -927,9 +927,9 @@ public class CUBRIDConnection implements Connection {
         return result;
     }
 
-    public synchronized int streamEnd() throws SQLException {
+    public synchronized long streamEnd() throws SQLException {
         checkIsOpen();
-        int result;
+        long result;
 
         synchronized (u_con) {
             result = u_con.streamEnd();
